♻️ refactor: 기존 통합 테스트 디렉토리 이전
